backend/src/requirements.txt
Django==2.2.7
psycopg2-binary==2.8.3

gunicorn==19.9.0

djangorestframework==3.10.3

pytest==5.1.2
pytest-django==3.5.1
pytest-factoryboy==2.0.3
pytest-cov==2.7.1

requests==2.22.0
dj-database-url==0.5.0

flake8==3.7.8

django-debug-toolbar==2.0.0
